HAI-1621 Disable endpoints that are not yet ready for production (#331)
Add a possibility to disable endpoints in certain environments.

Boolean environment variable HAITATON_FEATURE_HANKE_EDITING can be used to control this. The value should be false (disabled) in production and staging. Default value is true (enabled), so there is no need to set this elsewhere.

+++ services/hanke-service/src/main/kotlin/fi/hel/haitaton/hanke/attachment/hanke/HankeAttachmentController.kt
@@ -1,5 +1,6 @@
 package fi.hel.haitaton.hanke.attachment.hanke
 
+import fi.hel.haitaton.hanke.EndpointDisabledException
 import fi.hel.haitaton.hanke.HankeNotFoundException
 import fi.hel.haitaton.hanke.HankeService
 import fi.hel.haitaton.hanke.attachment.common.AttachmentInvalidException
@@ -18,6 +19,7 @@ import io.swagger.v3.oas.annotations.responses.ApiResponse
 import io.swagger.v3.oas.annotations.responses.ApiResponses
 import java.util.UUID
 import mu.KotlinLogging
+import org.springframework.beans.factory.annotation.Value
 import org.springframework.http.ResponseEntity
 import org.springframework.web.bind.annotation.DeleteMapping
 import org.springframework.web.bind.annotation.GetMapping
@@ -36,6 +38,7 @@ class HankeAttachmentController(
     private val hankeAttachmentService: HankeAttachmentService,
     private val hankeService: HankeService,
     private val permissionService: PermissionService,
+    @Value("\${haitaton.feature.hanke-editing}") val enableEditFeature: Boolean = true,
 ) {
 
     @GetMapping
@@ -120,6 +123,10 @@ class HankeAttachmentController(
         @PathVariable hankeTunnus: String,
         @RequestParam("liite") attachment: MultipartFile
     ): HankeAttachmentMetadata {
+        if (!enableEditFeature) {
+            throw EndpointDisabledException()
+        }
+
         permissionOrThrow(hankeTunnus, EDIT)
         return hankeAttachmentService.addAttachment(hankeTunnus, attachment)
     }
@@ -142,6 +149,9 @@ class HankeAttachmentController(
             ]
     )
     fun deleteAttachment(@PathVariable hankeTunnus: String, @PathVariable attachmentId: UUID) {
+        if (!enableEditFeature) {
+            throw EndpointDisabledException()
+        }
         permissionOrThrow(hankeTunnus, EDIT)
         return hankeAttachmentService.deleteAttachment(hankeTunnus, attachmentId)
     }
